requirements.txt
discord.py==2.6.4
dotenv==0.9.9
